+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next version
 
+- Add `payment_methods#list` endpoint
+
 ## v1.1.3
 
 - Add `payments#reschedule` endpoint